# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  types.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  types.UserFollowerResponse:
    properties:
      followers_ids:
        items:
          type: integer
        type: array
    type: object
  types.UserFollowingResponse:
    properties:
      followings_ids:
        items:
          type: integer
        type: array
    type: object
  types.UserPostsResponse:
    properties:
      posts_ids:
        items:
          type: integer
        type: array
    type: object
info:
  contact:
    email: maxuanquang@gmail.com
    name: Quang Ma
    url: https://www.linkedin.com/in/maxuanquang
  description: A simple social network management service API in Go using Gin framework
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://maxuanquang.dev/gin-social-network-tos
  title: Gin Social Network Service
  version: "1.0"
paths:
  /friends/{user_id}:
    delete:
      consumes:
      - application/json
      description: unfollow user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: unfollow user
      tags:
      - friends
    post:
      consumes:
      - application/json
      description: follow user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: follow user
      tags:
      - friends
  /friends/{user_id}/followers:
    get:
      consumes:
      - application/json
      description: get followers IDs of an user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UserFollowerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: get followers IDs of an user
      tags:
      - friends
  /friends/{user_id}/followings:
    get:
      consumes:
      - application/json
      description: get followings IDs of an user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UserFollowingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: get followings IDs of an user
      tags:
      - friends
  /friends/{user_id}/posts:
    get:
      consumes:
      - application/json
      description: get all posts user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.UserPostsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.MessageResponse'
      summary: get all posts of user
      tags:
      - friends
...
securityDefinitions:
  OAuth2Password:
    flow: password
    scopes:
      admin: "\t\t\t\t\t\t\tGrants read and write access to administrative information"
      read: "\t\t\t\t\t\t\t\tGrants read access"
      write: "\t\t\t\t\t\t\tGrants write access"
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"